docker/requirements.txt
pyjks==20.0.0
shortuuid==1.0.13
behavex==4.1.0
docker==7.1.0
confluent-kafka==2.8.0
PyYAML==6.0.2
m2crypto==0.41.0
watchdog==6.0.0
pyopenssl==25.0.0
azure-storage-blob==12.24.1
prometheus-api-client==0.5.5
humanfriendly==10.0
requests<2.29  # https://github.com/docker/docker-py/issues/3113
couchbase==4.3.5
